cvm-vsock: link a VSOCK to each TCP socket
To match the security requirements of CVM, VSOCK should be used in the
data path and the TCP/IP network stack should be excluded from it.

This patch introduce modofication that backs each TCP socket with a
VSOCK socket which listens/connects to the same port for data
transmission.

TODO: Currently this patch disables the normal functionality of the TCP
sockets, unless the corresponding userspace program is running on the
host.